fix: move key of advanced piano to separate ext. class
